Moves methods into PwgSession extensions to fix compiler issues and improve code organisation
