+++ piwigoKit/Network/pwg.images/PwgSession+Images.swift
@@ -1,0 +1,96 @@
+//
+//  PwgSession+Images.swift
+//  piwigoKit
+//
+//  Created by Eddy Lelièvre-Berna on 27/06/2023.
+//  Copyright © 2023 Piwigo.org. All rights reserved.
+//
+
+import os
+import Foundation
+
+public extension PwgSession {
+    
+    func getIDofImage(withMD5 md5sum: String,
+                      completion: @escaping (Int64?) -> Void,
+                      failure: @escaping (Error?) -> Void) {
+        // Launch request
+        let paramDict: [String : Any] = ["md5sum_list": md5sum]
+        postRequest(withMethod: pwgImagesExist, paramDict: paramDict,
+                    jsonObjectClientExpectsToReceive: ImagesExistJSON.self,
+                    countOfBytesClientExpectsToReceive: pwgImagesExistBytes) { jsonData in
+            do {
+                // Decode the JSON into codable type CommunityUploadCompletedJSON.
+                let decoder = JSONDecoder()
+                let uploadJSON = try decoder.decode(ImagesExistJSON.self, from: jsonData)
+
+                // Piwigo error?
+                if uploadJSON.errorCode != 0 {
+                    // Will retry later
+                    let error = self.localizedError(for: uploadJSON.errorCode,
+                                                    errorMessage: uploadJSON.errorMessage)
+                    failure(error)
+                    return
+                }
+
+                if let imageID = uploadJSON.data.first(where: {$0.md5sum == md5sum})?.imageID {
+                    completion(imageID)
+                } else {
+                    completion(nil)
+                }
+            }
+            catch {
+                let error = error as NSError
+                failure(error)
+                return
+            }
+        } failure: { error in
+            /// - Network communication errors
+            /// - Returned JSON data is empty
+            /// - Cannot decode data returned by Piwigo server
+            failure(error)
+        }
+    }
+
+    func setInfos(with paramsDict: [String: Any],
+                  completion: @escaping () -> Void,
+                  failure: @escaping (NSError) -> Void) {
+        postRequest(withMethod: pwgImagesSetInfo, paramDict: paramsDict,
+                    jsonObjectClientExpectsToReceive: ImagesSetInfoJSON.self,
+                    countOfBytesClientExpectsToReceive: pwgImagesSetInfoBytes) { jsonData in
+            // Decode the JSON object and check if image data were updated on server.
+            do {
+                // Decode the JSON into codable type ImagesSetInfoJSON.
+                let decoder = JSONDecoder()
+                let uploadJSON = try decoder.decode(ImagesSetInfoJSON.self, from: jsonData)
+                
+                // Piwigo error?
+                if uploadJSON.errorCode != 0 {
+                    let error = self.localizedError(for: uploadJSON.errorCode,
+                                                    errorMessage: uploadJSON.errorMessage)
+                    failure(error as NSError)
+                    return
+                }
+                
+                // Successful?
+                if uploadJSON.success {
+                    // Image properties successfully updated ▶ update image
+                    completion()
+                }
+                else {
+                    // Could not set image parameters
+                    failure(JsonError.unexpectedError as NSError)
+                }
+            } catch {
+                // Data cannot be digested
+                let error = error as NSError
+                failure(error)
+            }
+        } failure: { error in
+            /// - Network communication errors
+            /// - Returned JSON data is empty
+            /// - Cannot decode data returned by Piwigo server
+            failure(error)
+        }
+    }
+}
